global state is finally being handled correctly - still need to implement loading and error dispatches - may pivot to skills and then debugging jira issues endpoint
