+++ app/javascript/components/routes/Skills.js
@@ -1,9 +1,23 @@
 import React from 'react'
+import { useAppContext } from '../AppContext'
 import PropTypes from 'prop-types'
 
-const Skills = ({ skills }) => {
+const Skills = ({ }) => {
+  const state = useAppContext()
+
   return (
-    <div className="text-center">Skills</div>
+    <>
+      <div className="text-center">Skills</div>
+      <ul>
+        {
+        state?.skills?.map((skill) => {
+          return (
+            <li key={skill.id}>{skill.name}</li>
+          )
+        })
+        }
+      </ul>
+    </>
   )
 }
